corrected some compiler warnings.
git-svn-id: http://svn.berlios.de/svnroot/repos/sems/branches/internal_sip_stack@656 8eb893ce-cfd4-0310-b710-fb5ebe64c474

+++ core/plug-in/sipctrl/hash_table.cpp
@@ -51,7 +51,7 @@ trans_bucket _trans_table[H_TABLE_ENTRIES];
 
 trans_bucket::trans_bucket()
 {
-    id = (unsigned int)(trans_bucket*)(this - _trans_table);
+    id = (unsigned long)(trans_bucket*)(this - _trans_table);
     pthread_mutex_init(&m,NULL);
 }
